[TIP] move EmptyPage into own module
