+++ x-pack/plugins/threat_intelligence/public/modules/empty_page/empty_page.tsx
